Refactor two factor authentication to use AuthenticationSequence.Continue(). Implement bluetooth scanning framework.
This commit changes the way second factor authentication is handled by replacing
the primary and secondary approach with the AuthenticationSequence approach, in
which the order does not matter. This does not change anything in terms of
user interaction or GUI design, in which cases there are still primary and
secondary methods. AuthSeq.Continue() and AuthSeq.Next.

This commit also sets the groundwork for future implementation of bluetooth
scanning, by creating the RefreshBTDeviceLists() method which, depending on the
state, refreshes the bluetooth lists with the corresponding BT devices. All that
is left for complete functionality is backend functionality that actually scans
for BT devices. Currently, this is being faked in getBTDevicesInRange().