Fixed an invalid state where there were no users with FULL permission
If there were multiple users in the database and all users with FULL permission
deleted themselves, then we might be left only with NONE and LIMITED users.

To solve this, if the last FULL user deletes themselves, they will be notified
that doing so will delete all other users in addition to themselves, thus emptying
the database.