Beautification and enhancement of functionality for Initializing state
Adjusted the Initializing state GUI to fit the screen better with larger
text and buttons and spread out components.

Added a status message when adding a new user. If the new user attempts
to add an authentication method (card) which is already in the database,
rather than silently rejecting the card, an error message will appear
below the "Cancel" and "Save" buttons. This error message will go away
when the user swipes a valid card (one which isn't in the database) or
if the user clicks anywhere on the screen.