UserManagement: loseFocus() on Add New User, Error for Duplicate Auth
Upon clicking "Add New User" button, loseFocus() is called to ensure the
focus is not on the "Add New User" button. This will allow the user to
swipe a new card without clicking elsewhere in the form to lose focus
from the button.

Added error message when attempting to add a duplicate authentication
method to a new user. If a new user is being added with an
authentication method (card) that already exists in the database, rather
than silently rejecting the card, an error message will appear. This
error message will go away when the user swipes a valid card (one which
isn't in the database) or if the user clicks anywhere on the screen.